+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## [v4.20.0](https://github.com/auth0/node-auth0/tree/v4.20.0) (2025-03-19)
+
+[Full Changelog](https://github.com/auth0/node-auth0/compare/v4.19.0...v4.20.0)
+
+**Added**
+
+- Add support for RAR when using CIBA [\#1088](https://github.com/auth0/node-auth0/pull/1088) ([frederikprijck](https://github.com/frederikprijck))
+- Get all custom domains paginate support [\#1089](https://github.com/auth0/node-auth0/pull/1089) ([kushalshit27](https://github.com/kushalshit27))
+
+## [v4.19.0](https://github.com/auth0/node-auth0/tree/v4.19.0) (2025-03-12)
+
+[Full Changelog](https://github.com/auth0/node-auth0/compare/v4.18.0...v4.19.0)
+
+**Added**
+
+- Added tenant, channel in Branding Phone Providers [\#1085](https://github.com/auth0/node-auth0/pull/1085) ([tanya732](https://github.com/tanya732))
+- Added byok support for node auth0 [\#1080](https://github.com/auth0/node-auth0/pull/1080) ([tanya732](https://github.com/tanya732))
+- feature/formsCustomText [\#1082](https://github.com/auth0/node-auth0/pull/1082) ([tusharpandey13](https://github.com/tusharpandey13))
+- Add Support For Extensibility as Custom Provider in BrandingManager [\#1079](https://github.com/auth0/node-auth0/pull/1079) ([tanya732](https://github.com/tanya732))
+
 ## [v4.18.0](https://github.com/auth0/node-auth0/tree/v4.18.0) (2025-01-28)
 
 [Full Changelog](https://github.com/auth0/node-auth0/compare/v4.17.0...v4.18.0)